ベースの切り替え（まだ間にあう）

+++ Src/Manager/Camera/CameraEvent/MoveCameraEvent/MoveCameraEvent.h
@@ -2,7 +2,7 @@
 
 #include "../CameraEventBase.h"
 
-#include "../../../../Object/Common/Transform/Transform.h"
+#include "../../../../Common/Vector3.h"
 
 class MoveCameraEvent : public CameraEventBase
 {
@@ -15,7 +15,7 @@ public:
     /// <param name="endAngle">終了角度</param>
     /// <param name="endFov">終了視野角</param>
     /// <param name="frame">移動フレーム数</param>
-    MoveCameraEvent(const Vector3& endPos, const Vector3& endAngle, float endFov, int frame);
+    MoveCameraEvent(const Vector3& endPos, const Vector3& endAngle, float endFov, float frame);
 
     /// <summary>
     /// 現在のカメラ情報から指定したカメラ情報分移動した情報から、現在のカメラ情報（位置、角度、視野角）まで指定されたフレーム数で移動する
@@ -25,7 +25,7 @@ public:
     /// <param name="startLocalAngle">開始相対角度</param>
     /// <param name="startLocalFov">開始相対視野角</param>
     /// <param name="frame">移動フレーム数</param>
-    MoveCameraEvent(const CameraBase& camera, const Vector3& endLocalPos, const Vector3& endLocalAngle, float endLocalFov, int frame);
+    MoveCameraEvent(const CameraBase& camera, const Vector3& endLocalPos, const Vector3& endLocalAngle, float endLocalFov, float frame);
 
     ~MoveCameraEvent() override = default;
 
@@ -50,8 +50,8 @@ private:
     const float endFov;
 
     // 時間
-    const int maxFrame;
-    int currentFrame;
+    const float maxFrame;
+    float currentFrame;
 
     // 補間
     float EaseInOut(float t);
